fix: resolve TeaVM compilation errors
- Move WindowBackend interface to lwjgl-stubs module (teavm-runtime depends on it)
- Move ImageDecodeBackend interface and holder to lwjgl-stubs module
- Add missing setCursorPos method to CanvasWindowBackend
- Fix uninitialized variable error in AdaptiveRenderDistance frame callback

+++ addons/teavm-runtime/src/main/java/top/steve3184/webmc/chunk/ChunkMemoryManager.java
@@ -254,8 +254,8 @@ public final class ChunkMemoryManager {
      * Update player position for distance calculations.
      */
     public void updatePlayerPosition(double x, double z) {
-        this.playerChunkX = ChunkPos.getX(x);
-        this.playerChunkZ = ChunkPos.getZ(z);
+        this.playerChunkX = ChunkPos.getX((long) x);
+        this.playerChunkZ = ChunkPos.getZ((long) z);
     }
 
     /**
